requirements.txt
python-dotenv==1.0.1
python-telegram-bot==13.11
urllib3==1.26.15
google-cloud-dialogflow==2.34.0
requests==2.32.3